[Feature]: Add used quota updating logic (#19)
* add resource-quota test step

* add resource quota used calculator

* add clusterresourcequota update status

* add update resource quota to the cleaner phase

* add runtime client to find team and findnamespace functions

* move quota-related functions to a separate file

* fix linter

* use function currying for updatequotastatus

* fixed duplicate errors

* use userQuota for the total resource calculations

+++ api/v1alpha1/types.go
@@ -9,7 +9,7 @@ type UserQuota struct {
 	// max number of objects the user can store
 	MaxObjects resource.Quantity `json:"maxObjects,omitempty"`
 	// max number of buckets the user can create
-	MaxBuckets int `json:"maxBuckets,omitempty"`
+	MaxBuckets int64 `json:"maxBuckets,omitempty"`
 }
 
 // +kubebuilder:validation:Pattern=^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$
